requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.2
Django==3.2.5
django-allauth==0.45.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
filelock==3.0.12
idna==3.2
itypes==1.2.0
Jinja2==3.0.1
MarkupSafe==2.0.1
oauthlib==3.1.1
platformdirs==2.1.0
pyaml==20.4.0
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.1.2
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.6
virtualenv==20.6.0
